+++ src/kits/interface/layouter/ComplexLayouter.cpp
@@ -53,7 +53,7 @@ public:
 	void InitFromSizes(int32* sizes)
 	{
 		fLocations[0] = 0;
-		for (int32 i = 0; i < fCount; i++) 
+		for (int32 i = 0; i < fCount; i++)
 			fLocations[i + 1] = fLocations[i] + sizes[i] + fSpacing;
 	}
 
@@ -168,7 +168,7 @@ ComplexLayouter::ComplexLayouter(int32 elementCount, float spacing)
 	  fSums(new(nothrow) SumItem[elementCount + 1]),
 	  fSumBackups(new(nothrow) SumItemBackup[elementCount + 1]),
 	  fOptimizer(new(nothrow) LayoutOptimizer(elementCount)),
-	  fUnlimited(B_SIZE_UNLIMITED / (elementCount == 0 ? 1 : elementCount)),
+	  fUnlimited((int32)B_SIZE_UNLIMITED / (elementCount == 0 ? 1 : elementCount)),
 	  fMinMaxValid(false),
 	  fOptimizerConstraintsAdded(false)
 {
